+++ apps/backend/src/modules/activities/activities.ts
@@ -1,5 +1,6 @@
 import { logger } from "~/libs/modules/logger/logger.js";
 import { activityLikeRepository } from "~/modules/activity-likes/activity-likes.js";
+import { notificationService } from "~/modules/notifications/notifications.js";
 import { sectionStatusService } from "~/modules/section-statuses/section-statuses.js";
 
 import { ActivityController } from "./activity.controller.js";
@@ -11,6 +12,7 @@ const activityRepository = new ActivityRepository(ActivityModel);
 const activityService = new ActivityService({
 	activityLikeRepository,
 	activityRepository,
+	notificationService,
 });
 const activityController = new ActivityController(
 	logger,
@@ -18,7 +20,7 @@ const activityController = new ActivityController(
 	sectionStatusService,
 );
 
-export { activityController };
+export { activityController, activityService };
 export { ActivityEntity } from "./activity.entity.js";
 export { ActivityModel } from "./activity.model.js";
 export { type ActivityService } from "./activity.service.js";