+++ apps/backend/src/modules/notifications/notification.service.ts
@@ -5,10 +5,15 @@ import {
 	SocketNamespace,
 	type SocketService,
 } from "~/libs/modules/socket/socket.js";
-import { type Service } from "~/libs/types/types.js";
+import { type Service, type ValueOf } from "~/libs/types/types.js";
 
-import { NotificationStatus } from "./libs/enums/enums.js";
+import {
+	type NotificationFilter,
+	NotificationStatus,
+	type NotificationType,
+} from "./libs/enums/enums.js";
 import { NotificationError } from "./libs/exceptions/exceptions.js";
+import { filterQueryParameterToNotificationType } from "./libs/maps/maps.js";
 import {
 	type AllNotificationsResponseDto,
 	type CreateNotificationRequestDto,
@@ -31,21 +36,14 @@ class NotificationService implements Service {
 		this.socketService = socketService;
 	}
 
-	public async checkHasUserUnreadNotifications(
-		userId: number,
-	): Promise<boolean> {
-		return await this.notificationRepository.checkHasUserUnreadNotifications(
-			userId,
-		);
-	}
-
 	public async create(
 		payload: CreateNotificationRequestDto,
 	): Promise<NotificationResponseDto> {
-		const { receiverUserId, type, userId } = payload;
+		const { actionId, receiverUserId, type, userId } = payload;
 
 		const createdNotification = await this.notificationRepository.create(
 			NotificationEntity.initializeNew({
+				actionId,
 				receiverUserId,
 				status: NotificationStatus.UNREAD,
 				type,
@@ -56,7 +54,7 @@ class NotificationService implements Service {
 		const notification = createdNotification.toObject();
 
 		this.socketService.emitMessage({
-			event: SocketEvent.NEW_NOTIFICATION,
+			event: SocketEvent.UPDATE_NOTIFICATION,
 			payload: notification,
 			receiversIds: [String(notification.userId), String(receiverUserId)],
 			targetNamespace: SocketNamespace.NOTIFICATIONS,
@@ -76,6 +74,34 @@ class NotificationService implements Service {
 		}
 
 		return await this.notificationRepository.delete(notificationId);
+	}
+
+	public async deleteByActionId(
+		actionId: number,
+		type: ValueOf<typeof NotificationType>,
+	): Promise<boolean> {
+		const notification = await this.notificationRepository.findByActionId(
+			actionId,
+			type,
+		);
+
+		if (!notification) {
+			throw new NotificationError({
+				message: ExceptionMessage.NOTIFICATION_NOT_FOUND,
+				status: HTTPCode.NOT_FOUND,
+			});
+		}
+
+		const notificationObject = notification.toObject();
+
+		this.socketService.emitMessage({
+			event: SocketEvent.UPDATE_NOTIFICATION,
+			payload: notification,
+			receiversIds: [String(notificationObject.receiverUserId)],
+			targetNamespace: SocketNamespace.NOTIFICATIONS,
+		});
+
+		return await this.notificationRepository.deleteByActionId(actionId, type);
 	}
 
 	public async find(notificationId: number): Promise<NotificationResponseDto> {
@@ -101,17 +127,28 @@ class NotificationService implements Service {
 
 	public async findAllByReceiverUserId(
 		receiverUserId: number,
+		notificationFilter: ValueOf<typeof NotificationFilter>,
 		search: string,
 	): Promise<AllNotificationsResponseDto> {
+		const notificationType =
+			filterQueryParameterToNotificationType[notificationFilter];
+
 		const userNotifications =
 			await this.notificationRepository.findAllByReceiverUserId(
 				receiverUserId,
 				search,
+				notificationType,
 			);
 
 		return {
 			items: userNotifications.map((notification) => notification.toObject()),
 		};
+	}
+
+	public async getUnreadNotificationsCount(userId: number): Promise<number> {
+		return await this.notificationRepository.getUnreadNotificationsCount(
+			userId,
+		);
 	}
 
 	public async setReadNotifications(
@@ -138,11 +175,12 @@ class NotificationService implements Service {
 			});
 		}
 
-		const { receiverUserId, status, type, userId } = payload;
+		const { actionId, receiverUserId, status, type, userId } = payload;
 
 		const updatedNotification = await this.notificationRepository.update(
 			notificationId,
 			NotificationEntity.initializeNew({
+				actionId,
 				receiverUserId,
 				status,
 				type,