+++ apps/frontend/src/libs/modules/store/middlewares/notifications-socket.middleware.ts
@@ -4,6 +4,7 @@ import {
 	isAction,
 } from "@reduxjs/toolkit";
 
+import { NotificationFilter } from "~/libs/enums/enums.js";
 import { SocketEvent, SocketNamespace } from "~/libs/modules/socket/socket.js";
 import { type ExtraArguments } from "~/libs/modules/store/store.js";
 import { type AppDispatch, type RootState } from "~/libs/types/types.js";
@@ -19,9 +20,14 @@ const notificationsSocket = ({
 	);
 
 	return ({ dispatch }) => {
-		notificationsSocketInstance.on(SocketEvent.NEW_NOTIFICATION, () => {
-			void dispatch(userNotificationsActions.getUserNotifications());
-			void dispatch(userNotificationsActions.checkHasUserUnreadNotifications());
+		notificationsSocketInstance.on(SocketEvent.UPDATE_NOTIFICATION, () => {
+			void dispatch(
+				userNotificationsActions.getUserNotifications({
+					search: "",
+					type: NotificationFilter.ALL,
+				}),
+			);
+			void dispatch(userNotificationsActions.getUnreadNotificationsCount());
 		});
 
 		return (next) => {