+++ apps/frontend/src/modules/chat-messages/slices/chat-messages.ts
@@ -1,9 +1,10 @@
-import { sendMessage } from "./actions.js";
+import { sendMessage, setReadChatMessages } from "./actions.js";
 import { actions } from "./chat-messages.slice.js";
 
 const allAction = {
 	...actions,
 	sendMessage,
+	setReadChatMessages,
 };
 
 export { reducer } from "./chat-messages.slice.js";