+++ apps/frontend/src/pages/overview/libs/helpers/get-default-vendors.helper.ts
@@ -1,11 +1,21 @@
-import { type VendorResponseDto } from "~/modules/vendors/vendors.js";
+import {
+	VendorKey,
+	type VendorResponseDto,
+} from "~/modules/vendors/vendors.js";
+
+const DEFAULT_VENDOR = VendorKey.UDEMY;
 
 const getDefaultVendors = (
 	vendors: VendorResponseDto[],
 ): {
 	[k in string]: boolean;
 } => {
-	return Object.fromEntries(vendors.map((vendor) => [vendor.key, true]));
+	return Object.fromEntries(
+		vendors.map((vendor) => [vendor.key, vendor.key === DEFAULT_VENDOR]) as [
+			string,
+			boolean,
+		][],
+	);
 };
 
 export { getDefaultVendors };