+++ packages/shared/src/modules/user-notifications/enums/user-notifications-api-path.enum.ts
@@ -1,7 +1,7 @@
 const UserNotificationsApiPath = {
-	CHECK_HAS_USER_UNREAD_NOTIFICATIONS: "/check-has-user-unread-notifications",
 	READ_NOTIFICATIONS: "/read-notifications",
 	ROOT: "/",
+	UNREAD_COUNT: "/unread-count",
 } as const;
 
 export { UserNotificationsApiPath };