+++ packages/shared/src/modules/user-notifications/validation-schemas/validation-schemas.ts
@@ -1,1 +1,2 @@
 export { readNotificationsRequest } from "./read-notifications-request.validation-schema.js";
+export { userNotificationQueryParameters } from "./user-notification-query-parameters.validation-schema.js";